chore: resolve PR comments + prove should not automatically keygen
