+++ book/src/writing-apps/prove.md
@@ -1,14 +1,30 @@
 # Generating Proofs
 
-By default, the `prove` command will both `build` and `keygen` before generating a proof. This looks like:
+Generating a proof using the CLI is simple - first generate a key, then generate your proof. Using command defaults, this looks like:
 
 ```bash
+cargo openvm keygen
 cargo openvm prove [app | stark | evm]
 ```
 
+## Key Generation
+
+The `keygen` command generates both an application proving and verification key.
+
+```bash
+cargo openvm keygen
+    --config <path_to_app_config>
+```
+
+Similarly to `build`, `run`, and `prove`, options `--manifest-path`, `--target-dir`, and `--output-dir` are provided.
+
+If `--config` is not specified, the command will search for `openvm.toml` in the manifest directory. If the file isn't found, a default configuration will be used.
+
+The proving and verification key will be written to `${target_dir}/openvm/` (and `--output-dir` if specified).
+
 ## Proof Generation
 
-The `prove` CLI command, at its core, uses the options below. `prove` also gets access to all of the options that `run` has (see [Running a Program](../writing-apps/run.md) for more information).
+The `prove` CLI command, at its core, uses the options below. `prove` gets access to all of the options that `run` has (see [Running a Program](../writing-apps/run.md) for more information).
 
 ```bash
 cargo openvm prove [app | stark | evm]
@@ -18,7 +34,9 @@ cargo openvm prove [app | stark | evm]
     --proof <path_to_proof_output>
 ```
 
-If `--app-pk` and/or `--exe` are not provided, the command will call `keygen` and/or `build` respectively before generating a proof.
+If `--app-pk` is not provided, the command will search for a proving key at `${target_dir}/openvm/app.pk`.
+
+If `--exe` is not provided, the command will call `build` before generating a proof.
 
 If your program doesn't require inputs, you can (and should) omit the `--input` flag.
 
@@ -31,19 +49,3 @@ The `app` subcommand generates an application-level proof, the `stark` command g
 > In order to run the `evm` subcommand, you must have previously called the costly `cargo openvm setup`, which requires very large amounts of computation and memory (~200 GB).
 
 See [EVM Proof Format](./verify.md#evm-proof-json-format) for details on the output format for `cargo openvm prove evm`.
-
-
-## Key Generation
-
-You may want to generate keys separately - the `keygen` command allows you to do this.
-
-```bash
-cargo openvm keygen
-    --config <path_to_app_config>
-```
-
-Similarly to `build`, `run`, and `prove`, options `--manifest-path`, `--target-dir`, and `--output-dir` are provided.
-
-If `--config` is not specified, the command will search for `openvm.toml` in the manifest directory. If the file isn't found, a default configuration will be used.
-
-The proving and verification key will be written to `${target_dir}/openvm/` (and `--output-dir` if specified).
