+++ crates/cli/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod commands;
 pub mod default;
-pub mod global;
 pub mod input;
 pub mod util;
